Add bestscores in resultpage and delete react simple keyboard
