+++ package.json
@@ -34,7 +34,6 @@
     "react-router-dom": "^6.2.2",
     "react-scripts": "5.0.0",
     "react-share": "^4.4.0",
-    "react-simple-keyboard": "^3.4.83",
     "react-toastify": "^8.2.0",
     "typescript": "^4.5.5",
     "typewriter-effect": "^2.18.2",
