+++ src/pages/ResultPage/ResultPage.tsx
@@ -6,6 +6,7 @@ import FingerStatistcs from "./FingerStatistics";
 import { KeyData } from "../../data/keyboardData";
 import useProfileContext from "../../hooks/useProfileContext";
 import Header from "../../components/Header";
+import ResultTable from "../DashboardPage/ResultTable";
 
 // 子コンポーネントで使用する予定のデータ（現状はダミー）とその型
 
@@ -26,9 +27,20 @@ const ResultPage: React.FC<Props> = ({ currGameData }) => {
     <div>
       <Header />
       <Box sx={{ display: "flex", flexDirection: "column", alignItems: "center" }}>
-        <Box sx={{ display: "flex", flexDirection: "row", justifyContent: "space-evenly" }}>
+        <Box
+          sx={{
+            display: "flex",
+            flexDirection: "row",
+            justifyContent: "space-between",
+            alignItems: "start",
+            width: "90%",
+          }}
+        >
           <SimpleResult language={userSettings.codeLang} accuracy={currGameData.accuracy} speed={currGameData.speed} />
-          {/* <BestScores data={bestScores} /> */}
+          <div style={{ width: "50%" }}>
+            <h2>BestScores</h2>
+            <ResultTable tableHeight={270} />
+          </div>
         </Box>
         <br />
         <hr style={{ width: "90%" }} />
